Now we can export OoniTest
