+++ include/measurement_kit/ooni.hpp
@@ -5,6 +5,7 @@
 #include <measurement_kit/ooni/dns_injection_test.hpp>
 #include <measurement_kit/ooni/error.hpp>
 #include <measurement_kit/ooni/http_invalid_request_line_test.hpp>
+#include <measurement_kit/ooni/ooni_test.hpp>
 #include <measurement_kit/ooni/tcp_connect_test.hpp>
 #include <measurement_kit/ooni/templates.hpp>
 #endif
