+++ test/ooni/tcp_connect_test.cpp
@@ -5,7 +5,6 @@
 #define CATCH_CONFIG_MAIN
 #include "src/ext/Catch/single_include/catch.hpp"
 
-#include "src/ooni/ooni_test_impl.hpp"
 #include <chrono>
 #include <iostream>
 #include <list>
@@ -53,7 +52,7 @@ TEST_CASE("Make sure that set_output_path() works") {
                         .set_input_file_path("test/fixtures/hosts.txt")
                         .set_output_file_path("foo.txt")
                         .create_test_();
-    auto ptr = static_cast<ooni::OoniTestImpl *>(instance.get());
+    auto ptr = static_cast<ooni::OoniTest *>(instance.get());
     REQUIRE(ptr->output_filepath == "foo.txt");
 }
